Revert "Merge remote-tracking branch 'origin/proto-interface' into proto-interface"
This reverts commit 68e35aae209a7423d1f5720b36842befb6057a25, reversing
changes made to 960a099ca247c024989401a87397f6acd7347e42.

+++ core/src/main/java/org/polypheny/db/type/entity/temporal/PolyTime.java
@@ -92,7 +92,7 @@ public class PolyTime extends PolyTemporal {
 
 
     public static PolyTime of( Time value ) {
-        return new PolyTime( value.getTime(), TimeUnit.MILLISECOND );
+        return new PolyTime( (int) value.getTime(), TimeUnit.MILLISECOND );
     }
